daemon: extract routing-rule reconcile from daemon_apply.go
Pure code-motion split of pkg/daemon/daemon_apply.go (#5661): the
services (ip-monitoring) reconcile, routing-rule (rib-group/next-table)
reconcile, and route-leak snapshot reconcile move verbatim into
daemon_apply_routing.go.

Moved declarations (byte-identical bodies, no rename, no logic change):
applyServicesReconcile, applyRoutingRules, reconcileRouteLeakSnapshot.
Same package, no caller changes; per-file imports pruned by goimports.
Build: go build ./pkg/daemon/... green.
